+++ components/svg/v2/users.tsx
@@ -1,0 +1,39 @@
+import { FC } from 'react';
+
+import { SVGProps } from '../svg.types';
+
+const User: FC<SVGProps> = ({ maxWidth, maxHeight, ...props }) => (
+  <svg
+    style={{ maxWidth, maxHeight }}
+    viewBox="0 0 20 20"
+    fill="none"
+    {...props}
+  >
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M6.875 5C4.97652 5 3.4375 6.53902 3.4375 8.4375C3.4375 10.336 4.97652 11.875 6.875 11.875C8.77348 11.875 10.3125 10.336 10.3125 8.4375C10.3125 6.53902 8.77348 5 6.875 5ZM2.1875 8.4375C2.1875 5.84867 4.28617 3.75 6.875 3.75C9.46383 3.75 11.5625 5.84867 11.5625 8.4375C11.5625 11.0263 9.46383 13.125 6.875 13.125C4.28617 13.125 2.1875 11.0263 2.1875 8.4375Z"
+      fill="currentColor"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M13.2402 3.75001L13.2421 3.75C14.4853 3.75 15.6776 4.24386 16.5567 5.12294C17.4357 6.00202 17.9296 7.1943 17.9296 8.4375C17.9296 9.68071 17.4357 10.873 16.5567 11.7521C15.6776 12.6311 14.4853 13.125 13.2421 13.125C12.8969 13.125 12.6171 12.8452 12.6171 12.5C12.6171 12.1548 12.8969 11.875 13.2421 11.875C14.1538 11.875 15.0281 11.5128 15.6728 10.8682C16.3174 10.2235 16.6796 9.34918 16.6796 8.4375C16.6796 7.52582 16.3174 6.65148 15.6728 6.00682C15.0283 5.36239 14.1544 5.00025 13.243 5C12.9259 5.00105 12.6102 5.04363 12.3041 5.12665C11.971 5.217 11.6277 5.02018 11.5373 4.68704C11.447 4.3539 11.6438 4.01059 11.9769 3.92023C12.3888 3.80853 12.8135 3.7513 13.2402 3.75001Z"
+      fill="currentColor"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M6.87505 13.125C5.87208 13.125 4.88384 13.3664 3.99381 13.8287C3.10377 14.2911 2.33808 14.9609 1.76142 15.7815C1.56295 16.0639 1.17312 16.132 0.890703 15.9335C0.608285 15.7351 0.540226 15.3452 0.738689 15.0628C1.43069 14.0781 2.34952 13.2743 3.41756 12.7195C4.4856 12.1647 5.67149 11.875 6.87505 11.875C8.07862 11.875 9.2645 12.1647 10.3325 12.7195C11.4006 13.2743 12.3194 14.0781 13.0114 15.0628C13.2099 15.3452 13.1418 15.7351 12.8594 15.9335C12.577 16.132 12.1872 16.0639 11.9887 15.7815C11.412 14.9609 10.6463 14.2911 9.7563 13.8287C8.86626 13.3664 7.87802 13.125 6.87505 13.125Z"
+      fill="currentColor"
+    />
+    <path
+      fillRule="evenodd"
+      clipRule="evenodd"
+      d="M16.1239 13.8278C15.2338 13.3654 14.2455 13.1244 13.2426 13.125C12.8974 13.1252 12.6174 12.8456 12.6172 12.5004C12.617 12.1552 12.8966 11.8752 13.2418 11.875C14.4456 11.8743 15.6318 12.1636 16.7001 12.7185C17.7683 13.2734 18.6872 14.0775 19.3787 15.0628C19.5771 15.3453 19.5088 15.7351 19.2263 15.9334C18.9437 16.1317 18.5539 16.0635 18.3556 15.7809C17.7794 14.96 17.0139 14.2901 16.1239 13.8278Z"
+      fill="currentColor"
+    />
+  </svg>
+);
+
+export default User;
