+++ views/dapp/v2/lsd/statistics/prices-market-card/index.tsx
@@ -2,7 +2,7 @@ import { Box, Theme, Typography, useTheme } from '@interest-protocol/ui-kit';
 import { useTranslations } from 'next-intl';
 import { FC } from 'react';
 
-import { SuiPriceMarketProps } from '../overview.type';
+import { SuiPriceMarketProps } from '../statistics.type';
 import PriceMarketChart from './price-market-chart';
 import SuiPriceInfo from './sui-price-info';
 import TrendInfo from './trend-info';
