+++ views/dapp/v2/lsd/statistics/prices-market-card/price-market-chart.tsx
@@ -2,7 +2,7 @@ import { Box } from '@interest-protocol/ui-kit';
 import { FC } from 'react';
 
 import StepsChart from '../../../components/charts/steps-chart';
-import { ChartProps } from '../overview.type';
+import { ChartProps } from '../statistics.type';
 
 const PriceMarketChart: FC<ChartProps> = ({ data }) => (
   <Box height="3.125rem" width={['100%', '100%', '100%', '40%']}>
