+++ views/dapp/v2/lsd/statistics/statistics.data.ts
@@ -1,0 +1,596 @@
+export const DATA_CHART = [
+  {
+    date: '2023-05-12T21:00:00.000Z',
+    amount: 291.0431693000808,
+    description: 'Fri, 12 May 2023 21:00:00 GMT',
+    day: '12/5',
+  },
+  {
+    date: '2023-05-12T22:00:00.000Z',
+    amount: 291.0431693000808,
+    description: 'Fri, 12 May 2023 22:00:00 GMT',
+    day: '12/5',
+  },
+  {
+    date: '2023-05-12T23:00:00.000Z',
+    amount: 291.0431693000808,
+    description: 'Fri, 12 May 2023 23:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T00:00:00.000Z',
+    amount: 291.0431693000808,
+    description: 'Sat, 13 May 2023 00:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T01:00:00.000Z',
+    amount: 295.0781179713566,
+    description: 'Sat, 13 May 2023 01:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T02:00:00.000Z',
+    amount: 295.0781179713566,
+    description: 'Sat, 13 May 2023 02:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T03:00:00.000Z',
+    amount: 295.0781179713566,
+    description: 'Sat, 13 May 2023 03:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T04:00:00.000Z',
+    amount: 295.12493428521566,
+    description: 'Sat, 13 May 2023 04:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T05:00:00.000Z',
+    amount: 295.12493428521566,
+    description: 'Sat, 13 May 2023 05:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T06:00:00.000Z',
+    amount: 295.12493428521566,
+    description: 'Sat, 13 May 2023 06:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T07:00:00.000Z',
+    amount: 295.15264589438874,
+    description: 'Sat, 13 May 2023 07:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T08:00:00.000Z',
+    amount: 295.15264589438874,
+    description: 'Sat, 13 May 2023 08:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T09:00:00.000Z',
+    amount: 295.15264589438874,
+    description: 'Sat, 13 May 2023 09:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T10:00:00.000Z',
+    amount: 297.2979518751909,
+    description: 'Sat, 13 May 2023 10:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T11:00:00.000Z',
+    amount: 297.2979518751909,
+    description: 'Sat, 13 May 2023 11:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T12:00:00.000Z',
+    amount: 331.5025142838667,
+    description: 'Sat, 13 May 2023 12:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T13:00:00.000Z',
+    amount: 331.4521220517424,
+    description: 'Sat, 13 May 2023 13:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T14:00:00.000Z',
+    amount: 331.4521220517424,
+    description: 'Sat, 13 May 2023 14:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T15:00:00.000Z',
+    amount: 331.45889459683815,
+    description: 'Sat, 13 May 2023 15:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T16:00:00.000Z',
+    amount: 331.45889459683815,
+    description: 'Sat, 13 May 2023 16:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T17:00:00.000Z',
+    amount: 331.47722611769785,
+    description: 'Sat, 13 May 2023 17:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T18:00:00.000Z',
+    amount: 331.4770341383876,
+    description: 'Sat, 13 May 2023 18:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T19:00:00.000Z',
+    amount: 332.6077736955448,
+    description: 'Sat, 13 May 2023 19:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T20:00:00.000Z',
+    amount: 343.21265355451544,
+    description: 'Sat, 13 May 2023 20:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T21:00:00.000Z',
+    amount: 390.7295480619236,
+    description: 'Sat, 13 May 2023 21:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T22:00:00.000Z',
+    amount: 401.49716993673286,
+    description: 'Sat, 13 May 2023 22:00:00 GMT',
+    day: '13/5',
+  },
+  {
+    date: '2023-05-13T23:00:00.000Z',
+    amount: 401.49716993673286,
+    description: 'Sat, 13 May 2023 23:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T00:00:00.000Z',
+    amount: 401.50172173336426,
+    description: 'Sun, 14 May 2023 00:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T01:00:00.000Z',
+    amount: 395.0318595602843,
+    description: 'Sun, 14 May 2023 01:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T02:00:00.000Z',
+    amount: 395.0318595602843,
+    description: 'Sun, 14 May 2023 02:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T03:00:00.000Z',
+    amount: 395.0318595602843,
+    description: 'Sun, 14 May 2023 03:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T04:00:00.000Z',
+    amount: 395.0318595602843,
+    description: 'Sun, 14 May 2023 04:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T05:00:00.000Z',
+    amount: 395.0318595602843,
+    description: 'Sun, 14 May 2023 05:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T06:00:00.000Z',
+    amount: 395.0318595602843,
+    description: 'Sun, 14 May 2023 06:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T07:00:00.000Z',
+    amount: 441.32404599932164,
+    description: 'Sun, 14 May 2023 07:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T08:00:00.000Z',
+    amount: 445.32350395468205,
+    description: 'Sun, 14 May 2023 08:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T09:00:00.000Z',
+    amount: 445.32350395468205,
+    description: 'Sun, 14 May 2023 09:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T10:00:00.000Z',
+    amount: 445.3465021369105,
+    description: 'Sun, 14 May 2023 10:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T11:00:00.000Z',
+    amount: 445.3465021369105,
+    description: 'Sun, 14 May 2023 11:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T12:00:00.000Z',
+    amount: 445.3465021369105,
+    description: 'Sun, 14 May 2023 12:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T13:00:00.000Z',
+    amount: 445.35232171342733,
+    description: 'Sun, 14 May 2023 13:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T14:00:00.000Z',
+    amount: 445.35232171342733,
+    description: 'Sun, 14 May 2023 14:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T15:00:00.000Z',
+    amount: 445.35232171342733,
+    description: 'Sun, 14 May 2023 15:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T16:00:00.000Z',
+    amount: 445.48413707925977,
+    description: 'Sun, 14 May 2023 16:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T17:00:00.000Z',
+    amount: 454.5309779898704,
+    description: 'Sun, 14 May 2023 17:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T18:00:00.000Z',
+    amount: 454.5359036433123,
+    description: 'Sun, 14 May 2023 18:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T19:00:00.000Z',
+    amount: 458.5354675757311,
+    description: 'Sun, 14 May 2023 19:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T20:00:00.000Z',
+    amount: 458.5354675757311,
+    description: 'Sun, 14 May 2023 20:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T21:00:00.000Z',
+    amount: 458.5354675757311,
+    description: 'Sun, 14 May 2023 21:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T22:00:00.000Z',
+    amount: 458.5354675757311,
+    description: 'Sun, 14 May 2023 22:00:00 GMT',
+    day: '14/5',
+  },
+  {
+    date: '2023-05-14T23:00:00.000Z',
+    amount: 428.2697507020515,
+    description: 'Sun, 14 May 2023 23:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T00:00:00.000Z',
+    amount: 428.2697507020515,
+    description: 'Mon, 15 May 2023 00:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T01:00:00.000Z',
+    amount: 435.60525282950874,
+    description: 'Mon, 15 May 2023 01:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T02:00:00.000Z',
+    amount: 435.60525282950874,
+    description: 'Mon, 15 May 2023 02:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T03:00:00.000Z',
+    amount: 435.60525282950874,
+    description: 'Mon, 15 May 2023 03:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T04:00:00.000Z',
+    amount: 435.60525282950874,
+    description: 'Mon, 15 May 2023 04:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T05:00:00.000Z',
+    amount: 435.60525282950874,
+    description: 'Mon, 15 May 2023 05:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T06:00:00.000Z',
+    amount: 435.60525282950874,
+    description: 'Mon, 15 May 2023 06:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T07:00:00.000Z',
+    amount: 435.7017681539779,
+    description: 'Mon, 15 May 2023 07:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T08:00:00.000Z',
+    amount: 435.7017681539779,
+    description: 'Mon, 15 May 2023 08:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T09:00:00.000Z',
+    amount: 435.7017681539779,
+    description: 'Mon, 15 May 2023 09:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T10:00:00.000Z',
+    amount: 432.91809630863656,
+    description: 'Mon, 15 May 2023 10:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T11:00:00.000Z',
+    amount: 432.91809630863656,
+    description: 'Mon, 15 May 2023 11:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T12:00:00.000Z',
+    amount: 435.17587023936665,
+    description: 'Mon, 15 May 2023 12:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T13:00:00.000Z',
+    amount: 435.17587023936665,
+    description: 'Mon, 15 May 2023 13:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T14:00:00.000Z',
+    amount: 435.17587023936665,
+    description: 'Mon, 15 May 2023 14:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T15:00:00.000Z',
+    amount: 435.17587023936665,
+    description: 'Mon, 15 May 2023 15:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T16:00:00.000Z',
+    amount: 435.17004126888617,
+    description: 'Mon, 15 May 2023 16:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T17:00:00.000Z',
+    amount: 551.2750505056117,
+    description: 'Mon, 15 May 2023 17:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T18:00:00.000Z',
+    amount: 1294.7335703247636,
+    description: 'Mon, 15 May 2023 18:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T19:00:00.000Z',
+    amount: 1355.220728671987,
+    description: 'Mon, 15 May 2023 19:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T20:00:00.000Z',
+    amount: 1398.3382389667304,
+    description: 'Mon, 15 May 2023 20:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T21:00:00.000Z',
+    amount: 1402.2893737513773,
+    description: 'Mon, 15 May 2023 21:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T22:00:00.000Z',
+    amount: 1426.9538217858444,
+    description: 'Mon, 15 May 2023 22:00:00 GMT',
+    day: '15/5',
+  },
+  {
+    date: '2023-05-15T23:00:00.000Z',
+    amount: 4111.674605227576,
+    description: 'Mon, 15 May 2023 23:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T00:00:00.000Z',
+    amount: 4154.459230033042,
+    description: 'Tue, 16 May 2023 00:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T01:00:00.000Z',
+    amount: 4301.577864610145,
+    description: 'Tue, 16 May 2023 01:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T02:00:00.000Z',
+    amount: 4301.512106217926,
+    description: 'Tue, 16 May 2023 02:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T03:00:00.000Z',
+    amount: 4304.900869128154,
+    description: 'Tue, 16 May 2023 03:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T04:00:00.000Z',
+    amount: 4310.997243105704,
+    description: 'Tue, 16 May 2023 04:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T05:00:00.000Z',
+    amount: 4311.410706383893,
+    description: 'Tue, 16 May 2023 05:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T06:00:00.000Z',
+    amount: 4311.407453219374,
+    description: 'Tue, 16 May 2023 06:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T07:00:00.000Z',
+    amount: 4311.434031962864,
+    description: 'Tue, 16 May 2023 07:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T08:00:00.000Z',
+    amount: 4374.36747239479,
+    description: 'Tue, 16 May 2023 08:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T09:00:00.000Z',
+    amount: 4425.579960138658,
+    description: 'Tue, 16 May 2023 09:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T10:00:00.000Z',
+    amount: 4427.670242522202,
+    description: 'Tue, 16 May 2023 10:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T11:00:00.000Z',
+    amount: 4687.916729103646,
+    description: 'Tue, 16 May 2023 11:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T12:00:00.000Z',
+    amount: 5075.949637040903,
+    description: 'Tue, 16 May 2023 12:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T13:00:00.000Z',
+    amount: 5104.42176215155,
+    description: 'Tue, 16 May 2023 13:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T14:00:00.000Z',
+    amount: 5109.097061516665,
+    description: 'Tue, 16 May 2023 14:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T15:00:00.000Z',
+    amount: 5117.721343050018,
+    description: 'Tue, 16 May 2023 15:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T16:00:00.000Z',
+    amount: 5216.75695928635,
+    description: 'Tue, 16 May 2023 16:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T17:00:00.000Z',
+    amount: 5329.592720655521,
+    description: 'Tue, 16 May 2023 17:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T18:00:00.000Z',
+    amount: 5336.850721241441,
+    description: 'Tue, 16 May 2023 18:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T19:00:00.000Z',
+    amount: 5363.434637233742,
+    description: 'Tue, 16 May 2023 19:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T20:00:00.000Z',
+    amount: 5604.460365531229,
+    description: 'Tue, 16 May 2023 20:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T21:00:00.000Z',
+    amount: 5604.46033392627,
+    description: 'Tue, 16 May 2023 21:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T22:00:00.000Z',
+    amount: 5611.162829647179,
+    description: 'Tue, 16 May 2023 22:00:00 GMT',
+    day: '16/5',
+  },
+  {
+    date: '2023-05-16T23:00:00.000Z',
+    amount: 5652.582190535784,
+    description: 'Tue, 16 May 2023 23:00:00 GMT',
+    day: '17/5',
+  },
+];
