+++ views/dapp/v2/lsd/statistics/stats/index.tsx
