+++ views/dapp/v2/lsd/statistics/stats/stats-derivated-wrapper.tsx
